+++ reactivestate-core/src/commonMain/kotlin/com/ensody/reactivestate/AutoRun.kt
@@ -44,14 +44,14 @@ public fun CoroutineLauncher.autoRun(
  * @param onChange Gets called when the observables change. If you provide a handler you have to
  * manually call [run].
  * @param flowTransformer How changes should be executed/collected. Defaults to [conflatedWorker].
- * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
+ * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.main`.
  * @param withLoading Tracks loading state for the (re-)computation. Defaults to [CoroutineLauncher.loading].
  * @param observer The callback which is used to track the observables.
  */
 public fun CoroutineLauncher.coAutoRun(
     onChange: CoAutoRunOnChangeCallback<Unit>? = null,
     flowTransformer: AutoRunFlowTransformer = { conflatedWorker(transform = it) },
-    dispatcher: CoroutineDispatcher = dispatchers.default,
+    dispatcher: CoroutineDispatcher = dispatchers.main,
     withLoading: MutableValueFlow<Int>? = loading,
     observer: CoAutoRunCallback<Unit>,
 ): CoAutoRunner<Unit> =
@@ -103,7 +103,7 @@ public fun CoroutineScope.autoRun(
  * @param onChange Gets called when the observables change. If you provide a handler you have to
  * manually call [run].
  * @param flowTransformer How changes should be executed/collected. Defaults to [conflatedWorker].
- * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
+ * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.main`.
  * @param withLoading Tracks loading state for the (re-)computation. Defaults to `launcher.loading`.
  * @param observer The callback which is used to track the observables.
  */
@@ -111,7 +111,7 @@ public fun CoroutineScope.coAutoRun(
     launcher: CoroutineLauncher = SimpleCoroutineLauncher(this),
     onChange: CoAutoRunOnChangeCallback<Unit>? = null,
     flowTransformer: AutoRunFlowTransformer = { conflatedWorker(transform = it) },
-    dispatcher: CoroutineDispatcher = dispatchers.default,
+    dispatcher: CoroutineDispatcher = dispatchers.main,
     withLoading: MutableValueFlow<Int>? = launcher.loading,
     observer: CoAutoRunCallback<Unit>,
 ): CoAutoRunner<Unit> =
@@ -268,7 +268,7 @@ public class AutoRunner<T>(
  * @param onChange Gets called when the observables change. Your onChange handler has to
  * manually call [run] at any point (e.g. asynchronously) to change the tracked observables.
  * @param flowTransformer How changes should be executed/collected. Defaults to [conflatedWorker].
- * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
+ * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.main`.
  * @param withLoading Tracks loading state for the (re-)computation. Defaults to [CoroutineLauncher.loading].
  * @param immediate Whether to start tracking in a background coroutine immediately.
  * @param observer The callback which is used to track the observables.
@@ -277,7 +277,7 @@ public class CoAutoRunner<T>(
     launcher: CoroutineLauncher,
     onChange: CoAutoRunOnChangeCallback<T>? = null,
     flowTransformer: AutoRunFlowTransformer = { conflatedWorker(transform = it) },
-    private val dispatcher: CoroutineDispatcher = dispatchers.default,
+    private val dispatcher: CoroutineDispatcher = dispatchers.main,
     override val withLoading: MutableValueFlow<Int>? = launcher.loading,
     immediate: Boolean = false,
     private val observer: CoAutoRunCallback<T>,