+++ reactivestate-core/src/commonMain/kotlin/com/ensody/reactivestate/StateFlowUtils.kt
@@ -1,6 +1,9 @@
 package com.ensody.reactivestate
 
+import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.onSubscription
 
 /**
  * Replaces the [MutableStateFlow.value] with [block]'s return value.
@@ -24,3 +27,19 @@ public fun <T> MutableStateFlow<T>.replace(block: T.() -> T): T {
     value = value.block()
     return previous
 }
+
+/**
+ * Similar to [MutableStateFlow.onSubscription] but returns a [MutableStateFlow].
+ */
+public fun <T> MutableStateFlow<T>.onStateSubscription(
+    block: suspend FlowCollector<T>.() -> Unit,
+): MutableStateFlow<T> =
+    onSubscription(block).stateOnDemand { value }.toMutable { this@onStateSubscription.value = it }
+
+/**
+ * Similar to [StateFlow.onSubscription] but returns a [StateFlow].
+ */
+public fun <T> StateFlow<T>.onStateSubscription(
+    block: suspend FlowCollector<T>.() -> Unit,
+): StateFlow<T> =
+    onSubscription(block).stateOnDemand { value }