+++ reactivestate/src/androidMain/kotlin/com/ensody/reactivestate/android/Derived.kt
@@ -41,7 +41,7 @@ public fun <T> ViewModel.derived(
  *                Defaults to [SharingStarted.Eagerly].
  * @param launcher The [CoroutineLauncher] to use.
  * @param flowTransformer How changes should be executed/collected. Defaults to [conflatedWorker].
- * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
+ * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.main`.
  * @param withLoading Tracks loading state for the (re-)computation. Defaults to [CoroutineLauncher.loading] if
  *                    this is a [CoroutineLauncher] or `null` otherwise.
  * @param observer The callback which is used to track the observables.
@@ -51,7 +51,7 @@ public fun <T> ViewModel.derived(
     started: SharingStarted = SharingStarted.Eagerly,
     launcher: CoroutineLauncher = if (this is CoroutineLauncher) this else SimpleCoroutineLauncher(viewModelScope),
     flowTransformer: AutoRunFlowTransformer = { conflatedWorker(transform = it) },
-    dispatcher: CoroutineDispatcher = dispatchers.default,
+    dispatcher: CoroutineDispatcher = dispatchers.main,
     withLoading: MutableValueFlow<Int>? = if (this is CoroutineLauncher) launcher.loading else null,
     observer: CoAutoRunCallback<T>,
 ): StateFlow<T> =
@@ -87,7 +87,7 @@ public fun <T> LifecycleOwner.derived(
  *                Defaults to [SharingStarted.Eagerly].
  * @param launcher The [CoroutineLauncher] to use.
  * @param flowTransformer How changes should be executed/collected. Defaults to [conflatedWorker].
- * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
+ * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.main`.
  * @param withLoading Tracks loading state for the (re-)computation. Defaults to [CoroutineLauncher.loading] if
  *                    this is a [CoroutineLauncher] or `null` otherwise.
  * @param observer The callback which is used to track the observables.
@@ -97,7 +97,7 @@ public fun <T> LifecycleOwner.derived(
     started: SharingStarted = SharingStarted.Eagerly,
     launcher: CoroutineLauncher = if (this is CoroutineLauncher) this else LifecycleCoroutineLauncher(this),
     flowTransformer: AutoRunFlowTransformer = { conflatedWorker(transform = it) },
-    dispatcher: CoroutineDispatcher = dispatchers.default,
+    dispatcher: CoroutineDispatcher = dispatchers.main,
     withLoading: MutableValueFlow<Int>? = if (this is CoroutineLauncher) launcher.loading else null,
     observer: CoAutoRunCallback<T>,
 ): StateFlow<T> =